Complete MU puzzle implementation
An area of improvement here is that `xIIIy->xUy` and `xUUy->xy`
could implement the same helper procedure that locates the ith
group of III or UU respectively and returns the values of `start`
and `result`. The only logic in the transformation procedures then
would be to reconstruct what `str` should be after.