Minimal progress
The tranpose of a matrix with `m = Z` (a matrix with 0 rows) should be
`[]`. As seen from the pattern match of `lem1`, Idris does not know how
to check that a matrix with `m = Z` and `n = (S k)` (a matrix with 0
rows but non-zero columns) is still `[]`.

I am currently stuck on how to work with Idris to help it see this. Or,
Idris is currently stuck with me on helping me see something
different!

I am commiting for my own documentation of the path I take in proving
this.