Complete proof of matrix transpose involution
After switching the implementation of `tranpose` to use `map` the base
case was trivial.

The inductive step I struggled with. I had the concepts there, and I
knew I had to use `cong` and `cong2`, but it took a similar Agda
solution for me to realize I needed to use `trans`. After realizing I
needed `trans`, the rest of the inductive step followed soon after.